feat: 챗봇 코어 모듈 리팩토링
- `langgraph_core`를 `chat_core`로 변경하여 모듈 구조를 개선했습니다.
- `profile_updater.py`를 삭제하고 관련 기능을 서비스 레이어로 통합하여 데이터 관리 로직을 중앙화했습니다.
- UI 로직을 단순화하고 `chatbot.py`의 핸들러를 인라인으로 구현했습니다.
- 노트북 파일을 추가하여 챗봇 기능 테스트 및 개발을 용이하게 했습니다.
- `requirements.txt`와 `.gitignore`를 업데이트했습니다.
