+++ .gitignore
@@ -44,12 +44,11 @@ test.py
 
 # INFO: personal settings files when committing, turn on to ignore these files
 
-mcp server
-.mcp.json  
-*.md
-*.toml
-.claude**
-.gemini**
-.ruff**
-my_app/chatbot/notebooks/
-my_app/ui/level_quiz/data/generated/
+# .mcp.json  
+# *.md
+# *.toml
+# .claude**
+# .gemini**
+# .ruff**
+# my_app/chatbot/notebooks/
+# my_app/ui/level_quiz/data/generated/
