+++ environment.yml
@@ -1,0 +1,164 @@
+name: firegen
+channels:
+  - conda-forge
+dependencies:
+  - _python_abi3_support=1.0=hd8ed1ab_2
+  - aiohappyeyeballs=2.6.1=pyhd8ed1ab_0
+  - aiohttp=3.12.15=py310h5f69134_0
+  - aiosignal=1.4.0=pyhd8ed1ab_0
+  - altair=5.5.0=pyhd8ed1ab_1
+  - appnope=0.1.4=pyhd8ed1ab_1
+  - arro3-core=0.5.1=py310hde4708a_0
+  - asttokens=3.0.0=pyhd8ed1ab_1
+  - attrs=25.3.0=pyh71513ae_0
+  - aws-c-auth=0.9.0=h9eee66f_19
+  - aws-c-cal=0.9.2=hd08b81e_1
+  - aws-c-common=0.12.4=h6caf38d_0
+  - aws-c-compression=0.3.1=habbe1e8_6
+  - aws-c-event-stream=0.5.5=hd1b68e1_3
+  - aws-c-http=0.10.4=h09a8a51_0
+  - aws-c-io=0.21.2=hc6344be_1
+  - aws-c-mqtt=0.13.3=h625c29d_3
+  - aws-c-s3=0.8.6=h6ded10d_2
+  - aws-c-sdkutils=0.2.4=habbe1e8_1
+  - aws-checksums=0.2.7=habbe1e8_2
+  - aws-crt-cpp=0.33.1=h54a40e1_2
+  - aws-sdk-cpp=1.11.606=ha924a42_1
+  - azure-core-cpp=1.16.0=ha1c5762_0
+  - azure-identity-cpp=1.12.0=hd83eed2_0
+  - azure-storage-blobs-cpp=12.14.0=he094cc7_1
+  - azure-storage-common-cpp=12.10.0=h12fd690_2
+  - azure-storage-files-datalake-cpp=12.12.0=h30213e0_3
+  - brotli-python=1.1.0=py310h853098b_3
+  - bzip2=1.0.8=h99b78c6_7
+  - c-ares=1.34.5=h5505292_0
+  - ca-certificates=2025.7.14=hbd8a1cb_0
+  - cffi=1.17.1=py310h497396d_0
+  - colorama=0.4.6=pyhd8ed1ab_1
+  - comm=0.2.3=pyhe01879c_0
+  - cpython=3.10.18=py310hd8ed1ab_0
+  - debugpy=1.8.15=py310hf6ed2a0_0
+  - decorator=5.2.1=pyhd8ed1ab_0
+  - distro=1.9.0=pyhd8ed1ab_1
+  - duckdb=1.3.2=hca7485f_0
+  - exceptiongroup=1.3.0=pyhd8ed1ab_0
+  - executing=2.2.0=pyhd8ed1ab_0
+  - frozenlist=1.7.0=py310ha18c8e3_0
+  - gflags=2.2.2=hf9b8971_1005
+  - glog=0.7.1=heb240a5_0
+  - h11=0.16.0=pyhd8ed1ab_0
+  - h2=4.2.0=pyhd8ed1ab_0
+  - hpack=4.1.0=pyhd8ed1ab_0
+  - httpcore=1.0.9=pyh29332c3_0
+  - httpx=0.28.1=pyhd8ed1ab_0
+  - hyperframe=6.1.0=pyhd8ed1ab_0
+  - icu=75.1=hfee45f7_0
+  - idna=3.10=pyhd8ed1ab_1
+  - importlib-metadata=8.7.0=pyhe01879c_1
+  - ipykernel=6.29.5=pyh57ce528_0
+  - ipython=8.37.0=pyh8f84b5b_0
+  - jedi=0.19.2=pyhd8ed1ab_1
+  - jinja2=3.1.6=pyhd8ed1ab_0
+  - jiter=0.10.0=py310hde4708a_0
+  - jsonschema=4.25.0=pyhe01879c_0
+  - jsonschema-specifications=2025.4.1=pyh29332c3_0
+  - jupyter_client=8.6.3=pyhd8ed1ab_1
+  - jupyter_core=5.8.1=pyh31011fe_0
+  - krb5=1.21.3=h237132a_0
+  - libabseil=20250512.1=cxx17_hd41c47c_0
+  - libarrow=21.0.0=h4561df7_0_cpu
+  - libarrow-acero=21.0.0=h926bc74_0_cpu
+  - libarrow-compute=21.0.0=hd5cd9ca_0_cpu
+  - libarrow-dataset=21.0.0=h926bc74_0_cpu
+  - libarrow-substrait=21.0.0=hb375905_0_cpu
+  - libbrotlicommon=1.1.0=h5505292_3
+  - libbrotlidec=1.1.0=h5505292_3
+  - libbrotlienc=1.1.0=h5505292_3
+  - libcrc32c=1.1.2=hbdafb3b_0
+  - libcurl=8.14.1=h73640d1_0
+  - libcxx=20.1.8=hf598326_1
+  - libedit=3.1.20250104=pl5321hafb1f1b_0
+  - libev=4.33=h93a5062_2
+  - libevent=2.1.12=h2757513_1
+  - libexpat=2.7.1=hec049ff_0
+  - libffi=3.4.6=h1da3d7d_1
+  - libgoogle-cloud=2.39.0=head0a95_0
+  - libgoogle-cloud-storage=2.39.0=hfa3a374_0
+  - libgrpc=1.73.1=hcdac78c_0
+  - libiconv=1.18=hfe07756_1
+  - liblzma=5.8.1=h39f12f2_2
+  - libnghttp2=1.64.0=h6d7220d_0
+  - libopentelemetry-cpp=1.21.0=he15edb5_1
+  - libopentelemetry-cpp-headers=1.21.0=hce30654_1
+  - libparquet=21.0.0=h3402b2e_0_cpu
+  - libprotobuf=6.31.1=h702a38d_1
+  - libre2-11=2025.07.22=hb7c0934_0
+  - libsodium=1.0.20=h99b78c6_0
+  - libsqlite=3.50.3=h4237e3c_1
+  - libssh2=1.11.1=h1590b86_0
+  - libthrift=0.22.0=h14a376c_1
+  - libutf8proc=2.10.0=h74a6958_0
+  - libxml2=2.13.8=h52572c6_0
+  - libzlib=1.3.1=h8359307_2
+  - lz4-c=1.10.0=h286801f_1
+  - markupsafe=3.0.2=py310hc74094e_1
+  - matplotlib-inline=0.1.7=pyhd8ed1ab_1
+  - multidict=6.6.3=py310hdf261b0_0
+  - narwhals=1.48.1=pyhe01879c_0
+  - nbformat=5.10.4=pyhd8ed1ab_1
+  - ncurses=6.5=h5e97a16_3
+  - nest-asyncio=1.6.0=pyhd8ed1ab_1
+  - nlohmann_json=3.12.0=ha1acc90_0
+  - openssl=3.5.1=h81ee809_0
+  - orc=2.1.3=h3bfa610_0
+  - packaging=25.0=pyh29332c3_1
+  - parso=0.8.4=pyhd8ed1ab_1
+  - pexpect=4.9.0=pyhd8ed1ab_1
+  - pickleshare=0.7.5=pyhd8ed1ab_1004
+  - pip=25.1.1=pyh8b19718_0
+  - platformdirs=4.3.8=pyhe01879c_0
+  - polars=1.31.0=default_h13af070_1
+  - polars-default=1.31.0=py39h31c57e4_1
+  - prometheus-cpp=1.3.0=h0967b3e_0
+  - prompt-toolkit=3.0.51=pyha770c72_0
+  - propcache=0.3.1=py310hc74094e_0
+  - psutil=7.0.0=py310h078409c_0
+  - ptyprocess=0.7.0=pyhd8ed1ab_1
+  - pure_eval=0.2.3=pyhd8ed1ab_1
+  - pyarrow-core=21.0.0=py310h4fdadd6_0_cpu
+  - pycparser=2.22=pyh29332c3_1
+  - pygments=2.19.2=pyhd8ed1ab_0
+  - pysocks=1.7.1=pyha55dd90_7
+  - python=3.10.18=h6cefb37_0_cpython
+  - python-dateutil=2.9.0.post0=pyhe01879c_2
+  - python-duckdb=1.3.2=py310h1af2607_0
+  - python-fastjsonschema=2.21.1=pyhd8ed1ab_0
+  - python-gil=3.10.18=hd8ed1ab_0
+  - python_abi=3.10=8_cp310
+  - pyzmq=27.0.0=py310h6d25ecc_0
+  - re2=2025.07.22=h52998f3_0
+  - readline=8.2=h1d1bf99_2
+  - referencing=0.36.2=pyh29332c3_0
+  - rpds-py=0.26.0=py310hb4f9fe2_0
+  - ruff=0.12.5=h575f11b_0
+  - setuptools=80.9.0=pyhff2d567_0
+  - six=1.17.0=pyhe01879c_1
+  - snappy=1.2.2=hd121638_0
+  - sniffio=1.3.1=pyhd8ed1ab_1
+  - sqlglot=27.2.0=pyhe01879c_0
+  - stack_data=0.6.3=pyhd8ed1ab_1
+  - tk=8.6.13=h892fb3f_2
+  - tornado=6.5.1=py310h078409c_0
+  - tqdm=4.67.1=pyhd8ed1ab_1
+  - traitlets=5.14.3=pyhd8ed1ab_1
+  - urllib3=2.5.0=pyhd8ed1ab_0
+  - vegafusion=2.0.2=py310h8b84749_0
+  - vl-convert-python=1.7.0=py310h493c2e1_1
+  - wcwidth=0.2.13=pyhd8ed1ab_1
+  - wheel=0.45.1=pyhd8ed1ab_1
+  - yarl=1.20.1=py310hc74094e_0
+  - zeromq=4.3.5=hc1bb282_7
+  - zipp=3.23.0=pyhd8ed1ab_0
+  - zlib=1.3.1=h8359307_2
+  - zstd=1.5.7=h6491c7d_2
+prefix: /Users/yopchi/miniforge3/envs/firegen
