+++ my_app/chatbot/chat_core/prompts/analysis_investment_emotions.yaml
@@ -1,0 +1,21 @@
+_type: few_shot
+prefix: |
+  You are a skilled investment analyst. Analyze {compacted_user_answer} and extract all the noun keywords that describe the user's **emotions about investing**.
+
+  The results must adhere to the following conditions:
+
+  1.  Only include the noun forms of words that express emotions.
+  2.  Place each keyword into a list.
+  3.  There should be no duplicate words in the list.
+  
+examples:
+  - compacted_user_answer: "유망한 투자처를 남들보다 먼저 발견하여 자산 성장을 추구하며, 보유 자산의 가치가 일시적으로 하락하더라도 장기적인 관점에서 크게 신경 쓰지 않고 새로운 성장 동력을 탐색하는 데 집중하는 투자자입니다."
+    output: ['자신감', '인내', '태연함', '침착함', '낙관', '무관심', '열망', '열정', '도전정신', '적극성', '경쟁심', '욕구']
+  - compacted_user_answer: "시장 붕괴나 금리 인상 같은 악재를 미리 파악하여 손실을 피하는 것을 최우선으로 생각하고, 자산 가치 하락 시에는 손실이 더 커질까 봐 불안해하며 계속해서 계좌를 확인하는 위험 회피형 투자자입니다."
+    output: ['불안', '걱정', '두려움', '경계심', '초조함', '안전욕구']
+suffix: "input: {compacted_user_answer}\\noutput:"
+example_prompt:
+  _type: prompt
+  template: |
+    input: {compacted_user_answer}
+    output: {output}
