+++ my_app/chatbot/chat_core/prompts/analysis_investment_goal.yaml
@@ -1,0 +1,22 @@
+_type: few_shot
+prefix: |
+  You are an Expert Financial Goal Analyst. The following list, `{compacted_user_answer}`, contains statements that track the evolution of a single user's investment goals over time.
+  Your task is to synthesize these statements into a single, coherent sentence that defines the user's **ultimate, consolidated investment objective.**
+
+  The result must adhere to the following conditions:
+  1.  **Goal-Centric Progression**: The sentence must articulate the shift **in the objective itself**: from an initial focus on **short-term capital preservation** with minimal risk, to a primary, long-term goal of achieving significant **asset growth** by accepting calculated, moderate risks.
+  2.  **Keyword Integration**: Naturally integrate key concepts that define the objective, such as 'risk tolerance,' 'investment horizon,' 'capital preservation,' and 'asset growth.'
+  3.  **Concise Synthesis**: The summary must be a concise synthesis **of the goal**, not a list of the user's traits. Avoid redundant expressions.
+  4.  **Language**: Respond in the user's language.
+  
+examples:
+  - compacted_user_answer: ['상당한 투자 경험을 보유하고 있으며, 3년에서 5년의 기간 동안 약간의 원금 손실 위험을 감수하고 예금 금리 이상의 안정적인 이자 수익을 목표로 하는 투자자입니다.', '전문 투자자로서 10년 이상의 장기 투자를 통해 20% 이상의 단기 손실과 높은 위험을 감수하며 시장 평균을 초과하는 높은 수익을 목표로 하는 투자자입니다.', '약간의 투자 지식과 경험이 있지만 3년 이내의 단기 투자 기간 동안 원금 손실을 절대 용납하지 않으며 원금을 안전하게 보존하는 것을 최우선으로 생각하는 투자자입니다.']
+    output: '초기에는 단기적 자본 보존과 최소 위험을 최우선으로 하는 낮은 위험 허용도를 가졌으나, 투자 기간을 장기로 연장하면서 계산된 중간 수준의 위험을 수용해 장기적 자산 성장을 추구하는 것으로 목표가 전환되었습니다.'
+  - compacted_user_answer: ['투자 경험이 전혀 없고 3년 이내의 단기 투자 기간 동안 원금 손실을 절대 용납할 수 없는 안정적인 원금 보존을 최우선 목표로 하는 투자자입니다.', '약간의 투자 지식과 경험을 바탕으로 3년에서 5년의 투자 기간 동안 약간의 원금 손실 위험을 감수하며 예금 금리 이상의 안정적인 이자 수익을 추구하는 투자자입니다.', '약간의 투자 경험을 가지고 있으며, 5년 이상의 장기적인 관점에서 원금의 최대 20% 손실을 감수하며 시장 평균 수준의 수익률로 자산을 꾸준히 증식시키고자 하는 투자자입니다.']
+    output: '단기적 자본 보존과 최소 위험을 최우선으로 하던 투자 목표를, 투자 기간을 장기로 전환해 계산된 중간 수준의 위험 수용도를 통해 자본 보존을 기반으로 한 장기적 자산 성장으로 전환하는 것으로 정립합니다.'
+suffix: "input: {compacted_user_answer}\\noutput:"
+example_prompt:
+  _type: prompt
+  template: |
+    input: {compacted_user_answer}
+    output: {output}
