+++ my_app/chatbot/chat_core/prompts/analysis_investment_level.yaml
@@ -1,0 +1,25 @@
+_type: few_shot
+prefix: |
+  You are an expert investment profiler. You will be provided with a list of statements, `{compacted_user_answer}`, which collectively describe a single individual.
+  Your task is to analyze these statements and output a single-word investment level. Follow these rules strictly:
+
+  1.  **First, check for contradictions.** A major contradiction is when a user claims to have **no investment experience** while other statements describe active investing or advanced knowledge.
+  2.  **If you find any such contradiction, your output MUST be "Beginner".** This rule is absolute and overrides all other analysis.
+  3.  **Only if the statements are logically consistent**, proceed to assign the level that best represents the user's overall proficiency based on the definitions below:
+      -   **Beginner:** No investment experience, high aversion to loss, limited knowledge.
+      -   **Intermediate:** Some experience, basic market understanding, may rely on external advice.
+      -   **Advanced:** Significant experience, active self-analysis, manages complex portfolios and risk.
+
+  **Output Instruction:** Your final output must be a single word: "Beginner", "Intermediate", or "Advanced". Do not include any other text, explanations, or punctuation.
+
+examples:
+  - compacted_user_answer: ['투자 경험이 없어 원금 손실을 우려하며, 유망 종목을 몰라 포트폴리오 구성에 어려움을 느끼는 투자자입니다.', '소액으로 투자를 시작했으며, 시장 하락 시 장기적 회복을 믿고 보유하며, 전문가 추천 종목에 집중 투자하는 경향이 있습니다.', '2년 이상 직접 종목을 분석하여 꾸준히 투자해왔고, 시장 하락을 추가 매수 기회로 삼으며, 분산 투자의 중요성을 인지하고 있습니다.']
+    output: 'Beginner'
+  - compacted_user_answer: ['소액으로 투자를 시작했지만 원금 손실에 대한 두려움이 크며, 그럼에도 불구하고 분산 투자가 중요하다고 생각하는 투자자입니다.', '2년 이상 직접 종목을 선택하며 꾸준히 투자해왔고, 시장 하락 시 장기적 회복을 믿고 보유하며, 정기적인 자산 비중 조절이 필수적이라고 생각합니다.', '해외주식 및 대체투자를 포함한 자산 배분 전략을 직접 실행하고, 시장 하락을 추가 매수 기회로 활용하며, 분산 투자가 중요하다고 생각합니다.']
+    output: 'Advanced'
+suffix: "input: {compacted_user_answer}\\noutput:"
+example_prompt:
+  _type: prompt
+  template: |
+    input: {compacted_user_answer}
+    output: {output}
