+++ my_app/chatbot/chat_core/prompts/greeting.yaml
@@ -1,0 +1,22 @@
+_type: few_shot
+prefix: |
+  Next sentences are examples of greeting messages for the user by different user profile status.
+  Generate a personalized greeting message for the user.
+  - If {user_name} is empty, generate a generic welcome message with a random nickname.
+  - nickname_suggestions: ["투자 꿈나무", "미래의 자산가", "스마트 투자자"]
+examples:
+  - profile_status: "onboarding"
+    user_name: "김투자"
+    output: "안녕하세요, 김투자님! 자산구조대입니다. 투자 여정을 함께 하게되어 기대됩니다. 잘 부탁드립니다."
+  - profile_status: "editing"
+    user_name: "이분석"
+    output: "이분석님, 다시 오셨네요! 중단하셨던 부분부터 다시 시작해서 프로필을 완성해 보겠습니다!"
+  - profile_status: "completed"
+    user_name: "박성공"
+    output: "다시 오셨네요, 박성공님! 자산구조대와 함께 오늘도 성공적인 투자 습관을 만들어 보겠습니다. 오늘도 잘 부탁드립니다."
+suffix: "input: {profile_status}, {user_name}\\noutput:"
+example_prompt:
+  _type: prompt
+  template: |
+    input: {profile_status}, {user_name}
+    output: {output}
