+++ requirements.txt
@@ -1,30 +1,39 @@
 annotated-types==0.7.0
 anthropic==0.60.0
+anyio==4.10.0
+appdirs==1.4.4
 async-timeout==4.0.3
 blinker==1.9.0
 cachetools==5.5.2
+certifi==2025.7.14
+charset-normalizer==3.4.3
 click==8.2.1
 deepl==1.22.0
 deprecation==2.1.0
 docutils==0.21.2
+exceptiongroup==1.3.0
 feedparser==6.0.11
 filetype==1.2.0
 geographiclib==2.0
 geopy==2.4.1
 gitdb==4.0.12
 GitPython==3.1.45
-google-ai-generativelanguage==0.6.18
-google-api-core==2.25.1
-google-api-python-client==2.179.0
-google-auth==2.40.3
-google-auth-httplib2==0.2.0
-google-genai==1.31.0
-google-generativeai==0.8.5
+google-ai-generativelanguage
+google-api-core
+google-api-python-client
+google-auth
+google-auth-httplib2
+google-genai
+google-generativeai
 googleapis-common-protos==1.70.0
 gotrue==2.12.3
 grpcio==1.74.0
 grpcio-status==1.71.2
+h11==0.16.0
+httpcore==1.0.9
 httplib2==0.22.0
+httpx==0.28.1
+idna==3.10
 itsdangerous==2.2.0
 joblib==1.5.1
 jsonpatch==1.33
@@ -37,10 +46,10 @@ langchain-google-genai==2.1.9
 langchain-openai==0.3.30
 langchain-teddynote==0.3.45
 langchain-text-splitters==0.3.9
-langgraph==0.5.4
+langgraph==0.6.6
 langgraph-checkpoint==2.1.1
-langgraph-prebuilt==0.5.2
-langgraph-sdk==0.1.74
+langgraph-prebuilt==0.6.4
+langgraph-sdk==0.2.4
 langsmith==0.4.8
 lz4==4.4.4
 Markdown==3.8.2
@@ -51,12 +60,7 @@ nltk==3.9.1
 numpy==1.26.4
 olefile==0.47
 openai==1.100.1
-MarkupSafe==3.0.2
-matplotlib-inline==0.1.7
-multitasking==0.0.12
-narwhals==1.48.1
-nest-asyncio==1.6.0
-orjson==3.11.0
+orjson==3.11.1
 ormsgpack==1.10.0
 pandas==2.3.1
 pdf2image==1.17.0
@@ -68,17 +72,30 @@ postgrest==1.1.1
 proto-plus==1.26.1
 protobuf==5.29.5
 protoc-gen-openapiv2==0.0.1
+pyarrow==21.0.0
+pyasn1==0.6.1
+pyasn1_modules==0.4.2
+pydantic==2.11.7
+pydantic_core==2.33.2
+pydeck==0.9.1
+pyee==11.1.1
+PyJWT==2.10.1
+pymdown-extensions==10.16
+pyparsing==3.2.3
+pyppeteer==2.0.0
 python-dotenv==1.1.1
 pytz==2025.2
 PyYAML==6.0.2
 rank-bm25==0.2.2
-realtime==2.6.0
+realtime==2.7.0
 regex==2025.7.34
+requests==2.32.5
 requests-toolbelt==1.0.0
 rich==14.1.0
 rsa==4.9.1
 sgmllib3k==1.0.0
 smmap==5.0.2
+sniffio==1.3.1
 SQLAlchemy==2.0.41
 starlette==0.47.2
 storage3==0.12.0
@@ -88,18 +105,18 @@ StrEnum==0.4.15
 supabase==2.17.0
 supafunc==0.10.1
 tavily-python==0.7.10
-tenacity==8.5.0
+tenacity==9.1.2
 tiktoken==0.9.0
 toml==0.10.2
 tomlkit==0.13.3
-types-requests==2.32.4.20250611
+types-requests==2.32.4.20250809
 typing-inspection==0.4.1
+typing_extensions==4.15.0
 tzdata==2025.2
 uritemplate==4.2.0
+urllib3
 uvicorn==0.35.0
 watchdog==6.0.0
-websockets==15.0.1
+websockets
 xxhash==3.5.0
 zstandard==0.23.0
-finance-datareader
-pandas_ta
